Avoid specializing std::formatter for std::tm (not a great idea after all)
